ai-news-summarizer/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.0
langgraph==0.2.28
langchain-openai==0.2.0
tavily-python==0.5.0
pydantic==2.9.2
